fix(rotation): bump the spotlight collection, not just the window
the intended rotation of showdown collections is a pane cycling through
a subsample of cadidate showdown collections so the plex library isn't
overwhelmed with 100+ 5-15 item collections:

===[--X--]==================== day 1    key: 5 in movies/ --X--
====[--X--]=================== day 2         X is spotlight
=====[--X--]================== day 3       '=' are candidate showdowns

what was actually happening:

===[--X--]==================== day 1
====[-X---]=================== day 2
=====[X----]================== day 3
======[--X--]================= day 3

      ^-- spotlight had constant position for three days
         ^-- then resets to center window, every three days

this caused the same collection to post for three days, then jump to a
new index, etc. this effectively shadowbans up two 66% of showdown
collections from every reaching the plex frontpage

now, we don't just advance the window index, but the spotlight index
too.